requirements.txt
numpy==2.4.2
scipy==1.17.1
scikit-learn==1.8.0
pandas==3.0.1
pillow==10.4.0
torch==2.5.1
torchvision==0.20.1
torchreid==0.2.5
tqdm==4.67.3
matplotlib-inline==0.2.1
PyYAML==6.0.3
umap-learn==0.5.11
pynndescent==0.6.0
joblib==1.5.3
